dom/power/test/test_bug957893.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test bug 957893 - Crash in WakeLock</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
 <script type="application/javascript">
  SimpleTest.waitForExplicitFinish();
  function test() {
    try {
      var wl = navigator.requestWakeLock('');
      ok(false, "RequestWakeLock throws an exception!");
    } catch(e) {
      ok(true, "RequestWakeLock throws an exception!");
    }
    info("Still alive!");
    SimpleTest.finish();
  }
  SpecialPowers.pushPrefEnv({"set": [["dom.wakelock.enabled", true]]}, test);
 </script>
</body>
</html>
